refactor: override auth in fixture to make tests independent
