+++ tests/test_contact.py
@@ -6,7 +6,7 @@ from core.dependencies import (
 from db import Contact, Address, Email, Phone, ThingContactAssociation
 from db.engine import session_ctx
 from main import app
-from tests import client, cleanup_post_test, cleanup_patch_test
+from tests import client, cleanup_post_test, cleanup_patch_test, override_authentication
 from schemas.contact import ValidateEmail, ValidatePhone
 
 import pytest
@@ -14,26 +14,20 @@ from pydantic import ValidationError
 import re
 
 
-def override_authentication(default=True):
-    """
-    Override the authentication dependency for testing purposes.
-    This allows all users to be considered authenticated.
-    """
+@pytest.fixture(scope="module", autouse=True)
+def override_authentication_dependency_fixture():
 
-    def closure():
-        print("Overriding authentication")
-        return default
+    app.dependency_overrides[amp_admin_function] = override_authentication(
+        default={"name": "foobar", "sub": "1234567890"}
+    )
+    app.dependency_overrides[amp_editor_function] = override_authentication(
+        default={"name": "foobar", "sub": "1234567890"}
+    )
+    app.dependency_overrides[amp_viewer_function] = override_authentication()
 
-    return closure
+    yield
 
-
-app.dependency_overrides[amp_admin_function] = override_authentication(
-    default={"name": "foobar", "sub": "1234567890"}
-)
-app.dependency_overrides[amp_editor_function] = override_authentication(
-    default={"name": "foobar", "sub": "1234567890"}
-)
-app.dependency_overrides[amp_viewer_function] = override_authentication()
+    app.dependency_overrides = {}
 
 
 # ============= module & function fixtures =======================================
